Change object deserialization strategy
Objects now have synthetic class objects in deserialization as well. "Class"
proto message now can have a synthetic class object proto within, if it's not
contained in any external class file, which is the case with objects.
"class_object_present" boolean field is deprecated and will be dropped, since
its value is equivalent to the presence of "class_object" field